+++ lib/helpers/form_deps.dart
@@ -19,12 +19,13 @@ class FormDeps {
   FormDeps(this.form);
 
   /// Creates a subscription builder for one or more control names.
-  DepSubscription when(Iterable<String> fieldNames) =>
-      DepSubscription._(this, fieldNames.toList(growable: false));
+  DepSubscription when(Iterable<String> fieldNames) => DepSubscription._(this, fieldNames.toList(growable: false));
 
   /// Detaches all listeners registered through this instance.
   void dispose() {
-    for (final d in _disposers) d();
+    for (final d in _disposers) {
+      d();
+    }
     _disposers.clear();
   }
 
@@ -44,19 +45,20 @@ class DepSubscription {
   ///
   /// - Debounce/throttle can be added later without changing call sites.
   /// - Optionally runs once immediately via [runNow].
-  DepSubscription run(
-      VoidCallback callback, {
-        bool runNow = false,
-      }) {
+  DepSubscription run(VoidCallback callback, {bool runNow = false}) {
     for (final name in _fields) {
-      if (!_registered.add(name)) continue;
+      try{
+        if (!_registered.add(name)) continue;
 
-      final control = _deps.form.control<dynamic>(name);
+        final control = _deps.form.nestedControl(name);
 
-      void handler() => callback();
+        void handler() => callback();
 
-      control.valueNotifier.addListener(handler);
-      _deps._track(() => control.valueNotifier.removeListener(handler));
+        control.addListener(handler);
+        _deps._track(() => control.removeListener(handler));
+      } on DartNgFormsException {
+        continue;
+      }
     }
 
     if (runNow) callback();
@@ -66,10 +68,7 @@ class DepSubscription {
   /// Runs [callback] with a snapshot of the current values of the dependencies.
   ///
   /// Useful when your side-effect depends on both A and B values.
-  DepSubscription runWithValues(
-      void Function(Map<String, dynamic> values) callback, {
-        bool runNow = false,
-      }) {
+  DepSubscription runWithValues(void Function(Map<String, dynamic> values) callback, {bool runNow = false}) {
     return run(() {
       final values = <String, dynamic>{};
       for (final n in _fields) {